+++ src/dfxvm/list.rs
@@ -1,20 +1,42 @@
+use crate::dfxvm::cli::ListOpts;
+use crate::dfxvm::manifest::Manifest;
 use crate::error::dfxvm::ListError;
+use crate::json::fetch_json;
 use crate::locations::Locations;
 use crate::settings::Settings;
 use itertools::Itertools;
+use reqwest::Url;
 use semver::Version;
 
-pub fn list(locations: &Locations) -> Result<(), ListError> {
+pub async fn list(opts: ListOpts, locations: &Locations) -> Result<(), ListError> {
     let settings = Settings::load_or_default(&locations.settings_path())?;
-    let default_version = settings.default_version;
+    if opts.remote {
+        let url = Url::parse(&settings.manifest_url())?;
 
-    for version in installed_versions(locations)? {
-        let default_indicator = if default_version.as_ref() == Some(&version) {
-            " (default)"
+        info!("fetching {url}");
+        let manifest = fetch_json::<Manifest>(&url).await?;
+
+        if manifest.versions.is_empty() {
+            println!("No versions available.");
         } else {
-            ""
-        };
-        println!("{}{}", version, default_indicator);
+            println!("Available versions:");
+            let count = std::cmp::min(opts.number, manifest.versions.len());
+            let versions = manifest.versions.iter().rev().take(count);
+            for version in versions {
+                println!("  {}", version.to_string());
+            }
+        }
+    } else {
+        let default_version = settings.default_version;
+
+        for version in installed_versions(locations)? {
+            let default_indicator = if default_version.as_ref() == Some(&version) {
+                " (default)"
+            } else {
+                ""
+            };
+            println!("{}{}", version, default_indicator);
+        }
     }
     Ok(())
 }