+++ src/error/dfxvm.rs
@@ -46,8 +46,14 @@ pub enum Error {
 
 #[derive(Error, Debug)]
 pub enum ListError {
+    #[error("failed to fetch versions")]
+    FetchVersions(#[from] FetchJsonDocError),
+
     #[error(transparent)]
     LoadJsonFile(#[from] LoadJsonFileError),
+
+    #[error("failed to parse manifest url")]
+    ParseManifestUrl(#[from] url::ParseError),
 
     #[error("failed to read directory {path}")]
     ReadDir {